#346 Cleanup + adding test route for notification rest controller
