+++ quolance-api/src/main/java/com/quolance/quolance_api/controllers/NotificationRestController.java
@@ -1,16 +1,15 @@
 package com.quolance.quolance_api.controllers;
 
 import com.quolance.quolance_api.dtos.websocket.NotificationResponseDto;
-import com.quolance.quolance_api.dtos.websocket.SendNotificationRequestDto;
 import com.quolance.quolance_api.entities.Notification;
 import com.quolance.quolance_api.entities.User;
 import com.quolance.quolance_api.services.entity_services.UserService;
+import com.quolance.quolance_api.services.websockets.NotificationService;
 import com.quolance.quolance_api.services.websockets.impl.NotificationMessageService;
 import com.quolance.quolance_api.util.SecurityUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -24,30 +23,7 @@ public class NotificationRestController {
 
     private final NotificationMessageService notificationMessageService;
     private final UserService userService; // Added UserService to fetch User entities
-
-    @PostMapping("/send")
-    @Operation(summary = "Send a notification to specific users")
-    @ApiResponses(value = {
-            @ApiResponse(responseCode = "200", description = "Notification sent successfully"),
-            @ApiResponse(responseCode = "400", description = "Validation error")
-    })
-    public ResponseEntity<Void> sendNotification(@Valid @RequestBody SendNotificationRequestDto request) {
-        User sender = SecurityUtil.getAuthenticatedUser();
-        request.getRecipientIds().forEach(recipientId -> {
-            // Fetch recipient User object
-            User recipient = userService.findById(recipientId)
-                    .orElseThrow(() -> new RuntimeException("Recipient not found with ID: " + recipientId));
-
-            Notification notification = new Notification();
-            notification.setSender(sender);
-            notification.setRecipient(recipient); // Set the recipient User object
-            notification.setMessage(request.getMessage());
-            notification.setTimestamp(java.time.LocalDateTime.now());
-            notification.setRead(false);
-        });
-
-        return ResponseEntity.ok().build();
-    }
+    private final NotificationService notificationService;
 
     @PatchMapping("/{id}/read")
     @Operation(summary = "Mark a notification as read")
@@ -81,4 +57,16 @@ public class NotificationRestController {
                 .toList();
         return ResponseEntity.ok(response);
     }
+
+    @PostMapping("/send-test-notification")
+    @Operation(
+            summary = "Send a test notification",
+            description = "Send a test notification to the authenticated user"
+    )
+    public ResponseEntity<String> sendTestNotification(@RequestBody String message) {
+        User client = SecurityUtil.getAuthenticatedUser();
+        notificationMessageService.sendNotificationToUser(client, client, message);
+        return ResponseEntity.ok("Test notification sent successfully");
+    }
+
 }
