+++ quolance-api/src/main/java/com/quolance/quolance_api/services/business_workflow/impl/ClientWorkflowServiceImpl.java
@@ -146,11 +146,6 @@ public class ClientWorkflowServiceImpl implements ClientWorkflowService {
         return new PageImpl<>(freelancerDtos, pageable, userPage.getTotalElements());
     }
 
-    @Override
-    public void sendTestNotification(User user, String message) {
-        notificationMessageService.sendNotificationToUser(user, user, message);
-    }
-
 }
